chore(typescript): centralize types
