+++ src/extensions/snapshot/extend_world.ts
@@ -2,7 +2,7 @@ import Registry from '../../core/registry.js'
 import snapshot from './extension.js'
 import * as cmdOptions from './cmd_options.js'
 import { IWorld } from '@cucumber/cucumber'
-import { SnapshotOptions } from './snapshot.js'
+import { SnapshotOptions } from '../../types.js'
 
 const extendWorld = (world: IWorld, options?: SnapshotOptions) => {
     options = { ...cmdOptions, ...options }
